scripts/.config/scripts/python/spotify/requirements.txt
dbus_python==1.2.18
